RTDirRead/RTDirReadEx: sizes are size_t not unsigned.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@14062 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/HostServices/SharedFolders/testcase/tstShflCase.cpp
@@ -125,7 +125,7 @@ int rtDirClose(PRTDIR pDir)
     return VINF_SUCCESS;
 }
 
-int rtDirReadEx(PRTDIR pDir, PRTDIRENTRYEX pDirEntry, unsigned *pcbDirEntry, RTFSOBJATTRADD enmAdditionalAttribs)
+int rtDirReadEx(PRTDIR pDir, PRTDIRENTRYEX pDirEntry, size_t *pcbDirEntry, RTFSOBJATTRADD enmAdditionalAttribs)
 {
     switch(iDirList)
     {
@@ -225,7 +225,7 @@ static int vbsfCorrectCasing(char *pszFullPath, char *pszStartComponent)
 
     for(;;)
     {
-        uint32_t cbDirEntrySize = cbDirEntry;
+        size_t cbDirEntrySize = cbDirEntry;
 
         rc = RTDirReadEx(hSearch, pDirEntry, &cbDirEntrySize, RTFSOBJATTRADD_NOTHING);
         if (rc == VERR_NO_MORE_FILES)
